Rename models and router files to explicitly state their purpose
