+++ app/models/__init__.py
@@ -1,4 +1,7 @@
 # All models defining DB Table schemas should be imported into this module this allows them to be
 # found by Alembic when auto-generating migrations.
-from .cases import Case  # noqa: F401
-from .users import User  # noqa: F401
+from .case_model import Case  # noqa: F401
+from .user_model import User  # noqa: F401
+from .category_model import Category  # noqa: F401
+
+__all__ = ["Case", "User", "Category"]
