+++ app/routers/__init__.py
@@ -1,0 +1,4 @@
+from .case_router import router as case_router
+from .security_router import router as security_router
+
+__all__ = ["case_router", "security_router"]
