+++ app/routers/case_router.py
@@ -2,9 +2,9 @@ from typing import Sequence
 
 from fastapi import APIRouter, HTTPException, Depends
 
-from app.models.cases import CaseRequest, Case
-from app.models.users import User
-from app.models.categories import Category
+from app.models.case_model import CaseRequest, Case
+from app.models.user_model import User
+from app.models.category_model import Category
 from sqlmodel import Session, select
 from app.db import get_session
 from app.auth.security import get_current_active_user
